% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_unused_levels.R
\name{has_unused_levels}
\alias{has_unused_levels}
\title{Test for unused factor levels}
\usage{
has_unused_levels(f)
}
\arguments{
\item{f}{factor}
}
\value{
Logical \code{TRUE} or \code{FALSE} value.
}
\description{
Logical test of whether a factor has unused levels.
}
\author{
Alexander Zwart <alec.zwart at csiro.au>
}

